Remove docs generation from SDK
