--- docs/templates/all_tables.md.go.tpl
@@ -1,5 +1,0 @@
-# Source Plugin: {{.PluginName}}
-## Tables
-{{- range $table := $.Tables }}
-{{- template "all_tables_entry.md.go.tpl" $table}}
-{{- end }}
