--- docs/templates/all_tables_entry.md.go.tpl
@@ -1,5 +1,0 @@
-
-{{. | indentToDepth}}- [{{.Name}}]({{.Name}}.md){{ if .IsIncremental}} (Incremental){{ end }}
-{{- range $index, $rel := .Relations}}
-{{- template "all_tables_entry.md.go.tpl" $rel}}
-{{- end}}
