--- docs/testdata/TestGeneratePluginDocs-Markdown-README.md
@@ -1,10 +1,0 @@
-# Source Plugin: test
-
-## Tables
-
-- [incremental_table](incremental_table.md) (Incremental)
-- [test_table](test_table.md)
-  - [relation_table](relation_table.md)
-    - [relation_relation_table_a](relation_relation_table_a.md)
-    - [relation_relation_table_b](relation_relation_table_b.md)
-  - [relation_table2](relation_table2.md)
