--- docs/testdata/TestGeneratePluginDocs-Markdown-relation_relation_table_a.md
@@ -1,17 +1,0 @@
-# Table: relation_relation_table_a
-
-This table shows data for Relation Relation Table A.
-
-Description for relational table's relation
-
-The composite primary key for this table is ().
-
-## Relations
-
-This table depends on [relation_table](relation_table.md).
-
-## Columns
-
-| Name          | Type          |
-| ------------- | ------------- |
-|string_col|`utf8`|
