--- docs/testdata/TestGeneratePluginDocs-Markdown-relation_table.md
@@ -1,21 +1,0 @@
-# Table: relation_table
-
-This table shows data for Relation Table.
-
-Description for relational table
-
-The composite primary key for this table is ().
-
-## Relations
-
-This table depends on [test_table](test_table.md).
-
-The following tables depend on relation_table:
-  - [relation_relation_table_a](relation_relation_table_a.md)
-  - [relation_relation_table_b](relation_relation_table_b.md)
-
-## Columns
-
-| Name          | Type          |
-| ------------- | ------------- |
-|string_col|`utf8`|
